debian/libpam0g.postinst: Handle systemd units
Debian policy now considers init scripts optional.  When booted with
systemd, use systemd facilities for determining whether a service is
enabled and restarting.  We might want to try restarting services that
are running even if disabled.  I don't think we did that in the init
script case, and I don't know a good way to determine which units
should be restarted.

* Update template description because the services are no longer
considered init scripts on most systems.  It wouldn't be worth the
translation churn alone, but we have another template change in the
same version.

* Override lintian warning regarding calling systemctl.

oops lintian