_pam_limits_EPERM_NOT_FATAL
setrlimit will sometimes return EPERM for example if you try to increase the
number of open files too much.  This is not something we want to consider
fatal.  This also happens if you use non-root and try to decrease a limit.
Running PAM as non-root is not so great.

Authors: ?

Upstream status: submitted in <20070830171918.GB30563@dario.dodds.net>

Gbp-Pq: Name 032_pam_limits_EPERM_NOT_FATAL