pam_limits: Defaults if kernel limits unavailable
When set_all is set in pam_limits control, if we are not on linux, or
if parsing kernel limits fails, set explicit defaults that are similar
to linux defaults.

This patch is not particularly important to Debian on linux now that
set_all is no longer the default; kept mainly for non-linux ports.

based on patch by Peter Paluch <peterp@frcatel.fri.utc.sk>

Gbp-Pq: Name pam_limits_fallback_defaults